work on authentication js

+++ Develop/public/js/groupsHandlers/deleteGroupHandler.js
@@ -1,26 +1,32 @@
 const deleteGroupHandler = async (event) => {
     event.preventDefulat();
 
+    const id = document.querySelector('#deleteBtn').attributes('data-group-id')
+    const group_admin = document.querySelector('#deleteBtn').attributes('data-group-admin')
     // const id = document.querySelector('#deleteBtn').attributes('id')
     // is this right?
 
-    try {
-        const response = await fetch('/groups', {
-            method: 'DELETE',
-            body: JSON.stringify({ id }),
-            headers: { 'Content-Type': 'application/json' },
-        });
+    if (!id || !group_admin) {
+        alert("Failed to delete group.")
+    } else {
 
-        if (response.ok) {
-            document.location.replace('/')
-        } else {
-            alert('Failed to delete group. Please try again later.');
+        try {
+            const response = await fetch('/groups', {
+                method: 'DELETE',
+                body: JSON.stringify({ id, group_admin }),
+                headers: { 'Content-Type': 'application/json' },
+            });
+
+            if (response.ok) {
+                document.location.replace('/')
+            } else {
+                alert('Failed to delete group. Please try again later.');
+            }
+
+        } catch (err) {
+            alert(`Something went wrong: ${error}`)
         }
-
-    } catch (err) {
-        alert(`Something went wrong: ${error}`)
     }
-
 }
 
 document.querySelector('#deleteBtn').addEventListener('click', deleteGroupHandler)
